refactor(userADM): move value objects to dedicated file
